Modul Aufwandzeit auf 15min steps angepasst

+++ lib/schema/moduleUsages.ts
@@ -28,7 +28,7 @@ export const moduleUsages = pgTable(
     passed: boolean("passed"),
     note: text("note"),
     targetDate: timestamp("targetDate"),
-    hoursSpent: integer("hoursSpent"),
+    hoursSpent: decimal("hoursSpent"),
   },
   (t) => ({
     userModule: uniqueIndex().on(t.userId, t.moduleId),
